+++ src/components/copiable/copiable.js
@@ -4,39 +4,29 @@ import CopyButton from '../copy-button';
 import Popover from '../popover';
 import OSKey from 'os-key';
 import select from 'select';
-import Clipboard from 'clipboard/dist/clipboard.min.js';
-import debounce from 'debounce';
 import getWindow from '../utils/get-window';
 
 const DISABLE_CLICK_TO_SELECT_THRESHOLD = 640;
 const FEEDBACK_TIME = 2000;
 
+/**
+ * Make some text easily copiable.
+ *
+ * Comes with a [CopyButton](#copybutton) that you can click to automatically
+ * copy the text.
+ *
+ * Also, on wider viewports a click on the text will automatically select it
+ * all, so you can easily copy with your favorite keyboard shortcut.
+ */
 export default class Copiable extends React.Component {
-  static propTypes = {
-    /** The text that will be displayed and copied. */
-    value: PropTypes.string.isRequired,
-    /** Styling for the component wrapper. */
-    themeWrapper: PropTypes.string,
-    /** Styling for the copy button's container. */
-    themeCopyButtonContainer: PropTypes.string,
-    /** Styling for the copy button. */
-    themeCopyButton: PropTypes.string,
-    /** Styling for the text readout. */
-    themeText: PropTypes.string,
-    /** Styling for the copy hint tooltip. */
-    themeTooltip: PropTypes.string
-  };
-  static defaultProps = {
-    themeWrapper: 'relative round z0 scroll-styled',
-    themeCopyButtonContainer: 'absolute z2 top right mr3 mt3 color-white',
-    themeCopyButton: 'btn btn--xs py3 px3 round',
-    themeText: 'color-gray bg-darken5 round-l txt-mono py6 px6 pr36 txt-s'
-  };
+  constructor(props) {
+    super(props);
 
-  state = {
-    copyTooltipActive: false,
-    disableClickToSelect: false
-  };
+    this.copyHintText = '';
+    this.state = {
+      copyTooltipActive: false
+    };
+  }
 
   showCopyTooltip = () => {
     this.setState({ copyTooltipActive: true }, () => {
@@ -47,36 +37,19 @@ export default class Copiable extends React.Component {
   };
 
   componentDidMount() {
-    const { handleWindowResize } = this;
-    handleWindowResize();
-    window.addEventListener('resize', handleWindowResize);
-    this.copyHintText = `${this.getCopyKeys(
-      getWindow().navigator.userAgent
-    )} to copy!`;
+    this.copyHintText = (
+      <span>
+        <span className="txt-kbd">
+          {getCopyKeys(getWindow().navigator.userAgent)}
+        </span>{' '}
+        to copy
+      </span>
+    );
   }
 
-  componentWillUnmount = () => {
-    const { handleWindowResize, revertTimer } = this;
-    if (revertTimer) {
-      clearTimeout(revertTimer);
-    }
-    window.removeEventListener('resize', handleWindowResize);
-  };
-
-  handleWindowResize = debounce(() => {
-    const w = this.getWindowWidth();
-    if (
-      w <= DISABLE_CLICK_TO_SELECT_THRESHOLD &&
-      !this.state.disableClickToSelect
-    ) {
-      this.setState({ disableClickToSelect: true });
-    } else if (
-      w > DISABLE_CLICK_TO_SELECT_THRESHOLD &&
-      this.state.disableClickToSelect
-    ) {
-      this.setState({ disableClickToSelect: false });
-    }
-  }, 300);
+  componentWillUnmount() {
+    clearTimeout(this.revertTimer);
+  }
 
   setTextEl = el => {
     this.textEl = el;
@@ -90,72 +63,75 @@ export default class Copiable extends React.Component {
     return getWindow().innerWidth;
   }
 
-  getCopyKeys = ua => {
-    const keys = OSKey(ua);
-    if (!keys) return;
-    if (keys.primaryMeta) {
-      // ⌘+C
-      return `${keys.meta.symbol}+C`;
-    } else {
-      // Ctrl+C
-      return `${keys.ctrl.abbreviation}+C`;
-    }
-  };
-
-  handleSelect = () => {
-    if (this.state.disableClickToSelect) return;
+  handleTextFocus = () => {
+    if (getWindow().innerWidth < DISABLE_CLICK_TO_SELECT_THRESHOLD) return;
     select(this.getTextEl());
     this.showCopyTooltip();
   };
 
-  render() {
-    const {
-      value,
-      themeWrapper,
-      themeCopyButton,
-      themeCopyButtonContainer,
-      themeText,
-      themeTooltip
-    } = this.props;
+  handleTextBlur = event => {
+    if (!this.textEl.contains(event.relatedTarget)) {
+      this.setState({ copyTooltipActive: false });
+    }
+  };
 
-    const quickSelect = (
-      <span ref={this.setTextEl} onClick={this.handleSelect}>
-        {this.props.value}
-      </span>
-    );
+  render() {
+    const { props } = this;
 
     const copyBtn = typeof document !== 'undefined' &&
-      Clipboard.isSupported() && (
-        <div className={themeCopyButtonContainer}>
-          <CopyButton
-            text={value}
-            iconButtonProps={{
-              themeButton: themeCopyButton
-            }}
-            textEl={this.textEl}
-          />
+      CopyButton.isCopySupported() && (
+        <div className="fr py6 px6">
+          <CopyButton text={props.value} textEl={this.textEl} block={true} />
         </div>
       );
 
     const copyHintPopover = this.state.copyTooltipActive && (
       <Popover
-        themePopover={themeTooltip}
         getAnchorElement={this.getTextEl}
         placement="top"
         alignment="center"
         hideWhenAnchorIsOffscreen={true}
-        accessibleTitle={this.copyHintText}
+        accessibleTitle="Copy the selected text"
+        padding="small"
       >
         <div className="txt-s">{this.copyHintText}</div>
       </Popover>
     );
 
     return (
-      <div className={themeWrapper}>
-        <div className={themeText}>{quickSelect}</div>
+      <div className="clearfix bg-darken5 round">
         {copyBtn}
+        <div
+          tabIndex="-1"
+          ref={this.setTextEl}
+          onFocus={this.handleTextFocus}
+          onBlur={this.handleTextBlur}
+          className="py6 px12 txt-mono txt-s"
+          data-test="copiable-text-el"
+        >
+          <div className="my3">{props.value}</div>
+        </div>
         {copyHintPopover}
       </div>
     );
+  }
+}
+
+Copiable.propTypes = {
+  /**
+   * The text that will be displayed and copied.
+   */
+  value: PropTypes.string.isRequired
+};
+
+function getCopyKeys(ua) {
+  const keys = OSKey(ua);
+  if (!keys) return;
+  if (keys.primaryMeta) {
+    // ⌘+C
+    return `${keys.meta.symbol}+C`;
+  } else {
+    // Ctrl+C
+    return `${keys.ctrl.abbreviation}+C`;
   }
 }